+++ tests/ui/auto_enum/args.rs
@@ -40,32 +40,8 @@ mod marker {
         }
     }
 
-    #[auto_enum(marker(f), marker(g), Iterator)] //~ ERROR duplicate `marker` argument
-    fn multiple_marker(x: usize) -> impl Iterator<Item = i32> {
-        match x {
-            0 => 1..=8,
-            _ => 0..2,
-        }
-    }
-
-    #[auto_enum(marker(), Iterator)] //~ ERROR unexpected end of input, expected identifier
-    fn empty_marker(x: usize) -> impl Iterator<Item = i32> {
-        match x {
-            0 => 1..=8,
-            _ => 0..2,
-        }
-    }
-
-    #[auto_enum(marker(f, g), Iterator)] //~ ERROR unexpected token
-    fn marker_multiple_ident_1(x: usize) -> impl Iterator<Item = i32> {
-        match x {
-            0 => 1..=8,
-            _ => 0..2,
-        }
-    }
-
-    #[auto_enum(marker(f t), Iterator)] //~ ERROR unexpected token
-    fn marker_multiple_ident_2(x: usize) -> impl Iterator<Item = i32> {
+    #[auto_enum(marker(f), Iterator)] //~ ERROR expected `,`
+    fn marker_removed_delimiter(x: usize) -> impl Iterator<Item = i32> {
         match x {
             0 => 1..=8,
             _ => 0..2,
@@ -73,7 +49,7 @@ mod marker {
     }
 
     #[auto_enum(marker = f, marker = g, Iterator)] //~ ERROR duplicate `marker` argument
-    fn multiple_marker_eq(x: usize) -> impl Iterator<Item = i32> {
+    fn multiple_marker(x: usize) -> impl Iterator<Item = i32> {
         match x {
             0 => 1..=8,
             _ => 0..2,
@@ -81,7 +57,7 @@ mod marker {
     }
 
     #[auto_enum(marker =, Iterator)] //~ ERROR expected identifier
-    fn empty_marker_eq(x: usize) -> impl Iterator<Item = i32> {
+    fn empty_marker(x: usize) -> impl Iterator<Item = i32> {
         match x {
             0 => 1..=8,
             _ => 0..2,
@@ -89,7 +65,7 @@ mod marker {
     }
 
     #[auto_enum(marker = f t, Iterator)] //~ ERROR expected `,`
-    fn marker_eq_multiple_ident(x: usize) -> impl Iterator<Item = i32> {
+    fn marker_multiple_ident(x: usize) -> impl Iterator<Item = i32> {
         match x {
             0 => 1..=8,
             _ => 0..2,